# specs/002-fullstack-todo-app/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Full-Stack Todo Web Application API
  description: API for managing users and their todo tasks, including authentication, CRUD operations, filtering, and sorting.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Local Development Server
tags:
  - name: Authentication
    description: User authentication related operations
  - name: Tasks
    description: User task management operations
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRegistration:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
          pattern: "^[a-zA-Z0-9]+$"
          description: Unique alphanumeric username
        email:
          type: string
          format: email
          description: Valid email address
        password:
          type: string
          minLength: 8
          description: Password with at least one uppercase, one lowercase, one number, one special character
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$"
    UserLogin:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
          pattern: "^[a-zA-Z0-9]+$"
        password:
          type: string
          minLength: 8
          description: Password for login
    AuthSuccess:
      type: object
      properties:
        access_token:
          type: string
          description: JWT Access Token
        token_type:
          type: string
          example: bearer
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in-progress, completed]
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in-progress, completed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Message:
      type: object
      properties:
        message:
          type: string
paths:
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccess'
        '400':
          description: Bad request (e.g., validation errors)
        '409':
          description: Conflict (e.g., username or email already exists)
  /auth/login:
    post:
      summary: Log in an existing user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: User logged in successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccess'
        '401':
          description: Unauthorized (e.g., invalid credentials)
  /tasks:
    get:
      summary: Get all tasks for the authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: filter
          schema:
            type: string
            enum: [all, todo, in-progress, completed]
            default: all
          description: Filter tasks by status
        - in: query
          name: sort
          schema:
            type: string
            enum: [title, created_at]
            default: created_at
          description: Sort tasks by title or creation date (created_at)
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
    post:
      summary: Create a new task for the authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
  /tasks/{task_id}:
    parameters:
      - in: path
        name: task_id
        schema:
          type: string
          format: uuid
        required: true
        description: The ID of the task
    put:
      summary: Update an existing task
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Task not found or not owned by user
    delete:
      summary: Delete an existing task
      tags:
        - Tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized
        '404':
          description: Task not found or not owned by user